implement and use an optional number type that combines integer and float
At the level of JS, we probably don't care to distinguish these, so combine
them to make types a little simpler.

In particular, 0 gets identified as an integer, so if a field is
sometimes 0 and sometimes 1.0, it gets the type "integer | float"
which is awkward. Another approach to this specific problem would
be to add a specific type for zero that would get eliminated when
or-combined with either integer or float.